requirements.txt
flask==3.0.2
gunicorn==21.2.0
loguru==0.7.2
requests==2.31.0
aiohttp==3.9.3
aiosqlite==0.20.0
beautifulsoup4==4.12.3
openai==1.13.3
langchain==0.1.10
chromadb==0.4.24
tiktoken==0.6.0
python-dotenv==1.0.1
PyJWT==2.8.0
werkzeug==3.0.1
flask-cors==4.0.0
diskcache==5.6.3
html2text==2024.2.26
langchain-openai==0.1.3
py3langid==0.2.2
zhipuai==2.0.1
pymupdf==1.24.2
mammoth==1.7.1
python-pptx==0.6.23
pandas==2.2.1
tabulate==0.9.0
aiofiles==23.2.1
tokenizers==0.15.2
onnxruntime==1.16.3
numpy==1.26.4
